spread "fix: (helm) - do not add owner references to resources that contain the Helm keep resource-policy annotation #4389" to ansible-operator
